agent_managerのディレクトリを移動
